% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{assign.events.to.clones.of.a.tree}
\alias{assign.events.to.clones.of.a.tree}
\title{Assign events to clones based on presence/absence of them across samples}
\usage{
assign.events.to.clones.of.a.tree(tree, events, samples, cutoff = 0)
}
\arguments{
\item{tree:}{a merged tree}

\item{events:}{a data frame containing events and presence/absence
status in each sample (required colums: "event", followed by sample
columns representing VAF, each in a column}

\item{samples:}{samples (equivalent to vaf.col.names when calling
infer.clonal.models}

\item{cutoff:}{VAF cutoff to determine presence/absence of an event
in a sample}
}
\description{
Many events can be clustered correctly due to VAF can not
be estimated (eg. indels, copy number, copy altered SNVs). This function
try a best guess of what clone should the events belong to. It will look
at the "sample" column of the merged tree data frame and match with the
samples that we see the event. The clone that have the max number of
samples matching will be chosen.
}

