% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{assign.events.to.clones}
\alias{assign.events.to.clones}
\title{Wrapper to assign events to clones in all merged trees
This function calls assign.events.to.clones.of.a.tree
on each merged tree in list x$matched$merged.trees}
\usage{
assign.events.to.clones(x, events, samples, cutoff = 0)
}
\arguments{
\item{x:}{output of infer.clonal.models}

\item{events:}{see assign.events.to.clones.of.a.tree}

\item{samples:}{see assign.events.to.clones.of.a.tree}

\item{cutoff:}{see assign.events.to.clones.of.a.tree}
}
\description{
Wrapper to assign events to clones in all merged trees
This function calls assign.events.to.clones.of.a.tree
on each merged tree in list x$matched$merged.trees
}
