% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{clone.vaf.diff}
\alias{clone.vaf.diff}
\title{Test if two clones have different VAFs
Deprecated!}
\usage{
clone.vaf.diff(clone1.vafs, clone2.vafs, p.value.cut = 0.05)
}
\arguments{
\item{clone1.vafs:}{VAFs of all variants in clone/cluster 1}

\item{clone2.vafs:}{VAFs of all variants in clone/cluster 2}

\item{p.value.cut:}{significant level (if the test produce p.value
less than or equal to p.val => significantly different)}
}
\description{
Test if the two clones/clusters have different VAFs, using a
Mann-Whitney U test
}

