% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{compare.clone.trees}
\alias{compare.clone.trees}
\title{Compare two clonal evolution trees}
\usage{
compare.clone.trees(v1, v2, compare.seeding.models = T)
}
\arguments{
\item{v1:}{clonal evolution tree data frame 1}

\item{v2:}{clonal evolution tree data frame 2}

\item{compare.seeding.models:}{if true, match all sample status
(presence/absence/founding) at all nodes to make sure clonal
seeding between samples are preserved.}
}
\description{
Compare two clonal evolution trees to see if they differ
assuming excluded nodes are already excluded, labels are ordered

Return F if they are different, T if they are the same
}

