% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{compare.clone.trees.removing.leaves}
\alias{compare.clone.trees.removing.leaves}
\title{Compare two merged clonal evolution trees}
\usage{
compare.clone.trees.removing.leaves(v1, v2, ignore.seeding = FALSE)
}
\arguments{
\item{v1:}{clonal evolution tree data frame 1}

\item{v2:}{clonal evolution tree data frame 1}
}
\description{
Compare two clonal evolution trees to see if they differ
and if they also differ if all termial node (leaves) are removed.
This is useful to determine the number of unique models ignoring
sample specific clones which often give too many models when they
are present and low frequency. This function return 0 if tree match
at leaves, 1 if not match at leave but match when leaves are removed
2 if not matching at internal node levels
}
