% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trees.r
\name{convert.clone.to.branch}
\alias{convert.clone.to.branch}
\title{Create a tree from the merged.tree data.frame in clonevol output}
\usage{
convert.clone.to.branch(t, branch.lens = NULL,
  merged.tree.node.annotation = "sample.with.nonzero.cell.frac.ci")
}
\arguments{
\item{t:}{merged.tree data frame from clonevol}

\item{branch.lens:}{named vector of branch lengths, named according
to the cluster/clone ID/name. If not provided, all branches will receive
a same length}
}
\value{
Returns the same data.frame with additional columns, including:
$branches --> symbols of the branches corresponding to the clone
$blengths --> branch lengths
}
\description{
Create a branch-based clonal evolution tree from tree data.frame
}

