% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.r
\name{convert.consensus.tree.clone.to.branch}
\alias{convert.consensus.tree.clone.to.branch}
\title{Create trees for all merged.trees in clonevol output}
\usage{
convert.consensus.tree.clone.to.branch(x, cluster.col = "cluster",
  branch.scale = "none")
}
\arguments{
\item{x:}{output of infer.clonal.models}

\item{cluster.col:}{cluster column name (used to count variants
in each cluster/clone, and scale tree branch length}

\item{branch.scale:}{values=c('none', 'sqrt', 'log2'), scale branch
length by corresponding transformations. Note, branch length will
be estimated by the number of variants in each cluster/clone}
}
\description{
Take all merged/concensus trees in clonal evolution object
output of infer.clonal.models function, and convert them to branch-based
clonal evolution tree by adding additional columns and branch length scale,
=> make them ready to plot.
}
\examples{
data(aml1)
y = aml1
y <- convert.consensus.tree.clone.to.branch(y, branch.scale = 'sqrt')

}
