% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol2fishplot.r
\name{createFishPlotObjects}
\alias{createFishPlotObjects}
\title{Create a list of fishplot objects that can then be called}
\usage{
createFishPlotObjects(results)
}
\description{
Create a list of fishplot objects that can then be called
by layoutClones, then fishPlot
}
