% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{cross.rule.score}
\alias{cross.rule.score}
\title{Apply cross rule to all clones in all matched models using p-value combination}
\usage{
cross.rule.score(x, meta.p.method = "fisher", exhaustive.mode = FALSE,
  rank = TRUE, boot = NULL)
}
\arguments{
\item{x:}{output of infer.clonal.models}

\item{meta.p.method:}{method for combining p-values across samples
values = c('fisher', 'z'), default = 'fisher'}

\item{exhaustive.mode:}{placeholder for exhaustive.mode, not implemented yet.}
}
\description{
For each model, each clone will receive a score that
is equal to the  combined p-value of the test that the CCF of the
clone is >= 0
}

