% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{determine.subclone}
\alias{determine.subclone}
\title{Determine which clones are subclone in a single sample, also determine what
clones are possible founder clones of the samples}
\usage{
determine.subclone(v, r)
}
\arguments{
\item{v:}{data frame of subclonal structure as output of enumerate.clones
v must have row.names = v$lab}

\item{r:}{label of the clone where it and its decendent clones will be
evaluated. This is often the root of the tree}
}
\description{
Determine which clones are subclone or founder clone or both
or none subclones are identified based on cellular
fractions of the ancestor clones. Eg. if a clone is a subclone, all of its
decendent clones are subclone. If a clone is not a subclone and has zero
cell frac, its only decendent clone is not a subclone
founder clones are identified as clones whose cell frac is non-zero and whose
parent has zero cell frac
}
