% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.package.r
\name{draw.branch}
\alias{draw.branch}
\title{Draw tree branch using polygon that allows for
choosing border style and fill}
\usage{
draw.branch(x0, y0, x1, y1, w = 1, border.color = "black",
  border.linetype = "solid", border.width = 0.5, fill.color = NULL, ...)
}
\description{
Draw tree branch using polygon that allows for
choosing border style and fill
}
\examples{
plot(x=c(0,30), y=c(0, 30));
\dontrun{
draw.branch(c(5, 20, 12, 5, 5, 20),
 c(5, 20, 10, 20, 25, 25),
 c(10, 12, 20, 12, 10, 20),
 c(12, 15, 5, 12, 25, 29),
 border.linetype=c(1, 2, 3, 4,1,2),
 border.width=c(1,2,1,2,2,2))
}
}
