% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{draw.sample.clones.all}
\alias{draw.sample.clones.all}
\title{Draw all enumerated clonal models for a single sample}
\usage{
draw.sample.clones.all(x, outPrefix, object.to.plot = "polygon")
}
\arguments{
\item{x:}{output from enumerate.clones()}
}
\description{
Draw all enumerated clonal models for a single sample
}

