% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{draw.sample.clones}
\alias{draw.sample.clones}
\title{Draw clonal structures/evolution of a single sample}
\usage{
draw.sample.clones(v, x = 2, y = 0, wid = 30, len = 8, label = NULL,
  text.size = 1, cell.frac.ci = F, top.title = NULL)
}
\arguments{
\item{v:}{clonal structure data frame (output of enumerate.clones)}
}
\description{
Draw clonal structure for a sample with single or multiple
clones/subclones using polygon and tree plots
}

