% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deprecated.r
\name{enumerate.clones.monoclonal}
\alias{enumerate.clones.monoclonal}
\title{(Deprecated)
Enumerate all possible clonal structures for a single sample using
mono-clonal model (ie. the tumor originated from a single cell)}
\usage{
enumerate.clones.monoclonal(v)
}
\arguments{
\item{v:}{a data frame that contains clonal structure pre-identified from
clonality analysis (most often by clustering variants using their variant
allele frequencies). This data frame has two columns:
   cluster: a number indicating the cluster of variants
   vaf: variant allele frequency estimated for the cluster

Example data frame:\cr
cluster vaf\cr
1       0.5\cr
2       0.3\cr
3       0.1\cr}
}
\description{
(Deprecated)
Enumerate all possible clonal structures for a single sample using
mono-clonal model (ie. the tumor originated from a single cell)
}

