% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{enumerate.clones}
\alias{enumerate.clones}
\title{Enumerate all possible clonal structures for a single sample, employing the
subclonal test}
\usage{
enumerate.clones (v, sample, variants=NULL, subclonal.test.method='bootstrap'
, boot=NULL, p.value.cutoff=0.1)
}
\arguments{
\item{v:}{a data frame output of make.clonal.data.frame function}
}
\description{
Enumerate all possible clonal structures for a single sample
using monoclonal (ie. the primary tumor is originated from a single
cancer cell) or polyclonal model (ie. the primary tumor can originate from
multi cancer cells)
}
\details{
This function return a list of data frames. Each data frame
represents a clonal structure, similar to the output format of
make.clonal.data.frame output but now have 'parent' column identified
indicating the parent clone, and other columns (free, occupied) which
can be used to calculate cellular fraction and plotting. The root clone
has parent = -1, clones that have VAF=0 will have parent = NA
}
\examples{
--
}

