% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{estimate.ccf}
\alias{estimate.ccf}
\title{Calculate CI of CCF, pvals, etc.}
\usage{
estimate.ccf(vx, sample, i, boot, min.cluster.vaf, alpha, t = NULL,
  sub.clusters = NULL)
}
\arguments{
\item{vx:}{clonal evolution of a sample data frame}

\item{sample:}{name of vaf.col}

\item{i:}{row of vx where clone needs to be estimated}

\item{boot:}{pregenerated bootstrap}

\item{min.cluster.vaf:}{minimum cluster vaf to consider non-zero}

\item{alpha:}{alpha of CI}

\item{t:}{if subclonal.test was run already, provide the return object, if NULL
subclonal.test will be run
Return vx with additionally annotated columns related to CCF}
}
\description{
Calculate CI of CCF, pvals, etc.
}
