% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{estimate.clone.vaf}
\alias{estimate.clone.vaf}
\title{Estimate VAFs of clones/clusters from clonality analysis result}
\usage{
estimate.clone.vaf(v, cluster.col.name, vaf.col.names, vaf.in.percent = TRUE,
  ref.count.col.names = NULL, var.count.col.names = NULL,
  depth.col.names = NULL)
}
\arguments{
\item{v:}{data frame containing variants' VAF and clustering}

\item{cluster.col.name:}{name of the column that hold the cluster ID}

\item{vaf.col.names:}{names of the columns that hold VAFs}
}
\description{
Estimate VAFs of clones/clusters by calculating the median of
the VAFs of all variants provided.
}

