% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{extract.mapped.events}
\alias{extract.mapped.events}
\title{Produce a data frame of events mapped to clone and associated genes}
\usage{
extract.mapped.events(x)
}
\arguments{
\item{x:}{Output of assign.events.to.clones}
}
\description{
Events are separated by a comma in the event column
of the tree data frame. They will be resplitted and merge with
event data frame to produce event to clone/cluster
}
