% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deprecated.r
\name{find.matched.models.1}
\alias{find.matched.models.1}
\title{(Deprecated)
Find matched models between samples
infer clonal evolution models, given all evolve from the 1st sample
TODO: This function does not care about met met matching. Obsolete!}
\usage{
find.matched.models.1(vv, samples)
}
\description{
(Deprecated)
Find matched models between samples
infer clonal evolution models, given all evolve from the 1st sample
TODO: This function does not care about met met matching. Obsolete!
}

