% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{find.matched.models}
\alias{find.matched.models}
\title{Find matched models between samples
infer clonal evolution models, given all evolve from the 1st sample}
\usage{
find.matched.models(vv, samples, sample.groups = NULL,
  merge.similar.samples = FALSE)
}
\arguments{
\item{merge.similar.samples:}{see merge.clone.trees}
}
\description{
Find clonal evolution models across samples that are
compatible, given the models for individual samples
}
