% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.r
\name{generate.boot.nonparametric}
\alias{generate.boot.nonparametric}
\title{Generate and calculate bootstrap means for all clusters
Depricated!}
\usage{
generate.boot.nonparametric(variants, cluster.col.name = "cluster",
  vaf.col.names = NULL, vaf.in.percent = TRUE, num.boots = 1000,
  zero.sample = NULL)
}
\arguments{
\item{variants:}{data frame of variants with cluster assignments and VAF
of samples. Columns are c('cluster', 'sample1', 'sample2', ....)}

\item{zero.sample:}{The sample of zero vaf (to use to compare with other
clusters to determine if the cluster should be considered zero VAF, and
not included in the models)}
}
\description{
Generate and calculate bootstrap means for all clusters
Depricated!
}
