% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generate.boot.r
\name{generate.boot}
\alias{generate.boot}
\title{Generate and calculate bootstrap means for all clusters}
\usage{
generate.boot(variants, cluster.col.name = "cluster",
  cluster.center.method = "mean", vaf.col.names = NULL,
  depth.col.names = NULL, vaf.in.percent = TRUE, num.boots = 1000,
  bootstrap.model = "non-parametric", num.bernoulli.trials = "mean",
  weighted = FALSE, zero.sample = NULL, random.seed = NULL)
}
\arguments{
\item{variants:}{data frame of variants with cluster assignments and VAF
of samples. Columns are c('cluster', 'sample1vaf', 'sample2vaf', ...,
'sample1depth', 'sample2depth', ....). If possible, reduce input file to
only necessary columns. For unweighted resampling, columns will be assumed
to be VAF columns if not explicitly specified.}

\item{cluster.col.name:}{name of column containing cluster information. Must
be specified. Default: 'cluster'.}

\item{vaf.col.names:}{names of columns containing VAFs for each sample.
Default NULL. If weighted=FALSE and no VAF columns are specified, every
column except the cluster column will be treated as VAF columns.}

\item{depth.col.names:}{names of columns containing depth for each sample.
Default NULL. If weighted=TRUE, depth column names must all be specified.}

\item{vaf.in.percent:}{If TRUE, VAFs will be converted to proportions
between 0 and 1. Default TRUE.}

\item{num.boots:}{Number of times to resample. Default 1000.}

\item{bootstrap.model:}{specifies the statistical model used in bootstrap
resampling. Model can be normal, normal-truncated, beta, binomial,
beta-binomial, or non-parametric. Default: 'non-parametric'.}

\item{num.bernoulli.trials:}{the number of Bernoulli trials to perform when
using the beta-binomial model. For weighted analysis, cluster 'mean' depth
is the default. Otherwise, the user should specify what depth to use as a
numerical scalar or vector of depths for each cluster. If a scalar, that
value will be used for all clusters in all samples. If a vector, the length
must match the number of clusters times the number of samples.}

\item{weighted:}{If TRUE, weights variants proportionally to read count.
If TRUE, VAF and depth cluster columns must be specified. Default: FALSE.}

\item{zero.sample:}{The sample of zero vaf (to use to compare with other
clusters to determine if the cluster should be considered zero VAF, and
not included in the models)}
}
\description{
Generate and calculate bootstrap means for all clusters
}

