% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{generate.cloud.of.cells}
\alias{generate.cloud.of.cells}
\title{Generate a cloud of circles to represent cell population}
\usage{
generate.cloud.of.cells(colors, maxiter = 1000)
}
\arguments{
\item{colors:}{matching colors of the cells. Length(colors) will be
used as the number of cells to generate}

\item{maxiter:}{number of iterations for the layout algorithm
of the packcircles package}
}
\description{
Generate a cloud of circles to represent cell population
}
