% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{generate.fill.points}
\alias{generate.fill.points}
\title{Generate fill points for bell/polygon plots}
\usage{
generate.fill.points(x, y, num.points = 50)
}
\description{
Generate fill points for bell/polygon plots
}

