% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol2fishplot.r
\name{generateFishplotInputs}
\alias{generateFishplotInputs}
\title{Generate fishplot ready data from clonevol clonal evolution models}
\usage{
generateFishplotInputs(results, rescale = T, samples = NULL)
}
\arguments{
\item{results:}{output from infer.clonal.models}

\item{rescale:}{rescale VAF such that no sum rule is violated (eg. parent
should have VAF >= sum of its chilren's VAF}

\item{samples:}{the names of the samples to be used in fishplot (this should
be a subset of vaf.col.names parameter provided to infer.clonal.models); The
order of samples provided will be prerseved in the fishplot}
}
\description{
Generate fishplot ready data from clonevol clonal evolution models
}

