% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trees.package.r
\name{germinate}
\alias{germinate}
\title{Germinate a seed.}
\usage{
germinate(x, angle = 15, trunk.width = 20, middle = "0", left = "1",
  right = "2", left2 = "3", right2 = "4", left3 = "5", right3 = "6",
  left4 = "7", right4 = "8", plot = FALSE, ...)
}
\arguments{
\item{x}{Either a \code{seed} object returned by \code{\link{seed}}, or a
named \code{list} containing: \describe{\item{\code{trunk.height}}{Exactly
how glorious will this tree be?} \item{\code{branches}}{A binary coded
vector of addresses of branches to be included. Branches can branch to the
left or the right from the end of the trunk, or from the end of another
branch included in \code{branches}. Elements of \code{branches} may only
contain the characters given in args \code{left} and \code{right}, and all
parent branches of each element must also be provided. E.g. if \code{left}
and \code{right} are \code{'0'} and \code{'1'}, respectively, then
\code{'0'} is the first branch to the left stemming from the top of the
trunk, while \code{'1'} is the first branch to the right, stemming from the
top of the trunk; \code{'01'} would be a branch forking to the right from
the end of the first branch that forked left off the trunk.}
\item{\code{lengths}}{A vector of branch lengths corresponding to the
elements of \code{branches}. Should be the same length as \code{branches}.}}}

\item{angle}{The angle of branches relative to their parent branch (or
relative to the trunk). This angle is negated for left-pointing branches.}

\item{trunk.width}{The line width of the trunk. Widths are then scaled down
for successive child branches, to a minimum of 1.}

\item{left}{The character used to represent left-turning branches in the
\code{branches} vector (default is \code{'0'}). Must not be \code{'Y'}.}

\item{right}{The character used to represent right-turning branches in the
\code{branches} vector (default is \code{'1'}). Must not be \code{'Y'}.}

\item{plot}{Should the tree be plotted? (logical).}

\item{...}{Further arguments to \code{plot.plant}.}
}
\value{
a \code{plant} object, which is a \code{data.frame} comprising branch
  addresses, depths, lengths, angles, base coordinates, and tip coordinates.
}
\description{
Observe the miracle of life as your seed germinates and spews forth a glorious
tree.
}
\seealso{
\code{\link{seed}} \code{\link{foliate}} \code{\link{prune}}
}

