% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{get.cell.frac.ci}
\alias{get.cell.frac.ci}
\title{Get cellular fraction confidence interval}
\usage{
get.cell.frac.ci(vi, include.p.value = TRUE, sep = " - ")
}
\arguments{
\item{vi:}{clonal evolution tree data frame}

\item{include.p.value:}{include confidence level}

\item{sep:}{separator for the two confidence limits in output string

,p=', 1-vi$p.value,}
}
\description{
Get cellular fraction confidence interval, and also determine
if it is greater than zero (eg. if it contains zero). This function return
a list with $cell.frac.ci = strings of cell.frac.ci, $is.zero.cell.frac =
tell if cell.frac.ci contains zero
}
