% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{get.clonevol.colors}
\alias{get.clonevol.colors}
\title{Get the hex string of the preset colors optimized for plotting both
polygon plots and mutation scatter plots, etc.
#b3b3b3',
#999999',
#bf812d',
#aaaa55',
#c51b7d',
#ef6548',
#33a02c', '#3f007d', '#1f78b4',
#ffcc00',
#fca27e', '#fb8072',}
\usage{
get.clonevol.colors(num.colors, strong.color = FALSE)
}
\description{
Get the hex string of the preset colors optimized for plotting both
polygon plots and mutation scatter plots, etc.
#b3b3b3',
#999999',
#bf812d',
#aaaa55',
#c51b7d',
#ef6548',
#33a02c', '#3f007d', '#1f78b4',
#ffcc00',
#fca27e', '#fb8072',
}
