% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{get.model.max.p.value}
\alias{get.model.max.p.value}
\title{Get the highest p.value of the test of Ho:ccf<0
This is the 2nd strategy to score a model}
\usage{
get.model.max.p.value(v)
}
\description{
Get the highest p.value of the test of Ho:ccf<0
This is the 2nd strategy to score a model
}

