% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{get.model.non.negative.ccf.prob}
\alias{get.model.non.negative.ccf.prob}
\title{Get probability of a model for a sample
as the product of individual prob that ccf of a clone
is non-negative
This is a the 1st strategy to score a model}
\usage{
get.model.non.negative.ccf.prob(v)
}
\description{
Get probability of a model for a sample
as the product of individual prob that ccf of a clone
is non-negative
This is a the 1st strategy to score a model
}
