% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{get.subclones.across.samples}
\alias{get.subclones.across.samples}
\title{Get all set of subclones for all clone across samples
together with p value}
\usage{
get.subclones.across.samples(x, matched.model.index)
}
\description{
Get all set of subclones for all clone across samples
together with p value
}

