% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{import.tree}
\alias{import.tree}
\title{Import trees to clonevol}
\usage{
import.tree(tree.file, variant.file)
}
\arguments{
\item{tree.file}{A TSV file that contains the tree with minimum 3 labeled columns:
clone, parent, sample.with.nonzero.cell.frac.ci. It can also have additional
useful columns such as color, and events}

\item{variant.file}{A TSV file containing the variants, with cluster column,
vaf.col.names, etc.}
}
\value{
The clonal evolution tree similar to the output of infer.clonal.models
function that can be used in multiple plotting functions as if it is the output
of infer.clonal.models.
}
\description{
Import trees (eg. those predicted by other methods) to clonevol
thus we can visualize them using clonevol visualization features
}
\examples{
\dontrun{
y = import.tree('trees.tsv', 'variants.tsv')
}

}
