% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{is.ancestor}
\alias{is.ancestor}
\title{Check if clone a is ancestor of clone b in the clonal evolution tree}
\usage{
x = is.ancestor(v, a, b)
}
\arguments{
\item{v:}{clonal structure data frame}

\item{a:}{label of the cluster/clone a}

\item{b:}{label of the cluster/clone a}
}
\description{
Check if clone a is ancestor of clone b in the clonal evolution tree
}

