% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{make.graph}
\alias{make.graph}
\title{Construct igraph object from clonal structures of a sample}
\usage{
make.graph(v, cell.frac.ci = T)
}
\description{
Construct igraph object from clonal structures of a sample
}

