% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{match.sample.clones}
\alias{match.sample.clones}
\title{Check if two clonal structures are compatible (one evolves to the other)}
\usage{
match.sample.clones(v1, v2)
}
\arguments{
\item{v1:}{first clonal structure data frame}

\item{v2:}{first clonal structure data frame}
}
\description{
Check if two clonal structures are compatible (one evolves to
the other); ie. if structure v1 evolves to v2, all nodes in v2 must have
the same parents as in v1. This function returns TRUE if the two clonal
structures are compatible, otherwise, return FALSE
}
\details{
--
}
\examples{
--
}

