% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{merge.all.matched.clone.trees}
\alias{merge.all.matched.clone.trees}
\title{Recreate merged trees for matched models, given output of infer.clonal.models}
\usage{
\method{merge}{all.matched.clone.trees}(x)
}
\description{
Recreate merged trees for matched models, given output of infer.clonal.models
}
