% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.r
\name{merge.variants.and.events}
\alias{merge.variants.and.events}
\title{Merge variants and mapped events onto the same data frame that
is more convenient for plotting fancy boxplot of variants with
driver events highlighted}
\usage{
\method{merge}{variants.and.events}(variants, mapped.events = NULL,
  cluster.col.name = "cluster", event.col.name = "event", vaf.col.names,
  cn.col.names = c(), loh.col.names = c(), other.col.names = c())
}
\arguments{
\item{variants:}{variant data frame}

\item{mapped.events:}{data frame of events mapped onto cluster/clone
that is output assign.events.to.clones()$events}
}
\description{
Merge variants and mapped.events data frame to a single
data frame for fancy boxplot. Generally the columns of the two data frames
should be named the same, including (cluster.col.name, vaf.col.names,
cn.col.names, loh.col.names, column to highlight (driver), cluster, event).
event column should match between the two and events in mapped.events will
be excluded from variants data frame before merging them to prevent duplicate
}
