% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{plot.cell.population}
\alias{plot.cell.population}
\title{Plot a tumor as a cloud of cells reflecting the cellular frequencies}
\usage{
\method{plot}{cell.population}(cell.frac, colors, labels = NULL,
  cell.cex = 2, delta = NULL, cell.border.color = "black",
  cell.border.size = 0.1, num.cells = 200, layout = "cloud",
  clone.grouping = "random", frame = FALSE)
}
\arguments{
\item{cell.frac:}{cellular fraction of the clones (0-1)}

\item{colors:}{colors of the clones}

\item{num.cells:}{number of cells to plot}

\item{layout:}{"plate" (square) or "cloud" (default = "cloud")}

\item{cell.border.color:}{see plot.cloud.of.cells}

\item{cell.border.size:}{see  plot.cloud.of.cells}

\item{clone.grouping:}{see  plot.cloud.of.cells}

\item{frame:}{put a frame around the cell cloud}
}
\description{
Plot a tumor as a cloud of cells reflecting the cellular frequencies
}
