% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{plot.clonal.models}
\alias{plot.clonal.models}
\title{Visualize clonal evolution models using various plots.}
\usage{
plot.clonal.models(x, out.prefix, ...)
}
\arguments{
\item{x:}{output of  infer.clonal.models function}

\item{out.dir:}{output directory for the plots}

\item{models.to.plot:}{row numbers of the models to plot in x$matched$index}

\item{scale.monoclonal.cell.frac:}{c(TRUE, FALSE); if TRUE, scale cellular
fraction in the plots (ie. cell fraction will be scaled by 1/purity =
1/max(VAF*2)), default = TRUE}

\item{width:}{width of the plots (in), if NULL, automatically choose width}

\item{height:}{height of the plots (in), if NULL, automatically choose height}

\item{out.format:}{format of the plot files ('png', 'pdf', 'pdf.multi.files')}

\item{resolution:}{resolution of the PNG plot file}

\item{overwrite.output:}{if TRUE, overwrite output directory, default=FALSE}

\item{max.num.models.to.plot:}{max number of models to plot; default = 10}

\item{individual.sample.tree.plot:}{c(TRUE, FALSE); plot individual sample trees
if TRUE, combined graph that preserved sample labels will be produced in graphml
output, default = FALSE}

\item{merged.tree.plot:}{Also plot the merged clonal evolution tree across
samples}

\item{merged.tree.node.annotation:}{the name of the column in
x$matched$mereged.trees[[i]] that will be used to annotate the clones in trees
see also parameter node.annotation of function plot.tree
default = 'sample.with.nonzero.cell.frac.ci'}

\item{merged.tree.cell.frac.ci:}{Show cell fraction CI for samples in merged tree}

\item{tree.node.label.split.character:}{sometimes the labels of samples are long,
so to display nicely many samples annotated at leaf nodes, this parameter
specify the character that splits sample names in merged clonal evolution
tree, so it will be replaced by line feed to display each sample in a line,}

\item{tree.node.num.samples.per.line:}{Number of samples displayed on each line
in the tree node; default NULL (do not attempt to distribute samples on lines)}

\item{trimmed.tree.plot:}{Also plot the trimmed clonal evolution trees across
samples in a separate PDF file}

\item{color.node.by.sample.group:}{color clones by grouping found in sample.group.
based on the grouping, clone will be stratified into different groups according
to what sample group has the clone signature variants detected. This is useful
when analyzing primary, metastasis, etc. samples and we want to color the clones
based on if it is primary only, metastasis only, or shared, etc. etc.}

\item{merged.tree.node.size.scale:}{scale merged tree node size by this value,
compared to individual tree node size}

\item{merged.tree.node.text.size.scale:}{scale merged tree node text size
by this value, compared to  individual tree text size}

\item{clone.time.step.scale:}{scaling factor for distance between the tips
of the polygon/bell representing clone}

\item{zero.cell.frac.clone.color:}{color clone with zero cell fraction
in the sample with this color (default = NULL, color using matching color}

\item{zero.cell.frac.clone.border.color:}{border color of the bell for clones that
are not found in sample; if equal "fill", the fill color of bell is used
auto-generated)}

\item{cell.frac.ci:}{Display cell frac CI}

\item{disable.cell.frac:}{Completely remove cell frac CI info. in plots}

\item{samples:}{samples to plot (ordered), equal vaf.col.names used in
infer.clonal.models}

\item{bell.border.width:}{border with of bell curve}

\item{merged.tree.clone.as.branch:}{default=TRUE, plot merged.tree with branch
as clone (mtcab)}

\item{mtcab.angle:}{mtcab branch angle in degree}

\item{mtcab.branch.width:}{mtcab branch witdh in points}

\item{mtcab.branch.text.size:}{mtcab branch text size (cex)}

\item{mtcab.node.size=3:}{mtcab node size}

\item{mtcab.node.label.size:}{mtcab node label text size}

\item{mtcab.node.text.size:}{mtcab node annotation text size}

\item{mtcab.event.sep.char:}{mtcab event separator (used to break down eventa}

\item{mtcab.show.event:}{show mapped events on branches
to multiple lines)}

\item{mtcab.tree.rotation.angle:}{c(0 - bottom up, 90 - left to right,
180 - top down); angle to rotate the tree, default = 180}

\item{mtcab.tree.text.angle:}{text angle, if NULL auto-determine}

\item{xstarts:}{sample-named vector of x axis start positions of the samples
which will be used to layout sample bell plots over clinical timeline}

\item{xstops:}{sample-named vector of x axis stop positions of the samples}

\item{fancy.variant.boxplot.*:}{map to corresponding params of variant.box.plot}

\item{merged.tree.distance.from.bottom:}{distance from the plot area for the
merged tree (inches). If the height of combined final plot is large, increase this
to make sure the tree is not stretched vertically}

\item{panel.widths:}{vector of widths of the panels to plot which will be used to
scale the widths of variant boxplot, bell plot, and tree against each other}

\item{trimmed.merged.tree.plot:}{plot trimmed merged tree (T/F, default=T)}

\item{trimmed.merged.tree.plot.width:}{width (inches)}

\item{trimmed.merged.tree.plot.height:}{height (inches)}
}
\description{
Plot evolution models and clonal admixtures inferred by
infer.clonal.models function using various plots, including: bell plots,
sphere of cells, node-based and branch-based clonal evolution trees.
}

