% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{plot.clonevol.colors}
\alias{plot.clonevol.colors}
\title{Plot ClonEvol colors}
\usage{
\method{plot}{clonevol.colors}(num.colors = 40)
}
\arguments{
\item{num.colors}{Number of colors}
}
\description{
Plot ClonEvol colors
}

