% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{plot.cloud.of.cells}
\alias{plot.cloud.of.cells}
\title{Plot a cloud of cells}
\usage{
\method{plot}{cloud.of.cells}(cells, title = "", alpha = 1, frame = FALSE,
  cell.border.color = "black", cell.border.size = 0.1,
  clone.grouping = "random", limits = c(-50, 50))
}
\arguments{
\item{cells:}{cell cloud as returned from generate.cloud.of.cells}

\item{frame:}{draw a frame surrounding the cloud of cells}

\item{cell.border.color:}{color of the border of the circles used
to draw a cell (default = black)}

\item{cell.border.size:}{line size of the cell border, the smaller
the figure size is, the smaller this value needs to be (default = 0.1)}

\item{clone.grouping:}{how the cells of the same clone being grouped
values are c("random", "horizontal", "vertical"), default="random"}
}
\description{
Plot a cloud of cells
}
