% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxplot.r
\name{plot.cluster.flow}
\alias{plot.cluster.flow}
\title{Plot the mean/median of the clusters of variants across samples}
\usage{
\method{plot}{cluster.flow}(var, cluster.col.name = "cluster",
  ignore.clusters = NULL, vaf.col.names = NULL, sample.names = NULL,
  vaf.in.percent = TRUE, center.measure = "median", low.vaf.no.line = F,
  min.cluster.vaf = 0, line.size = 1, shape.size = 5, colors = NULL,
  shapes = NULL, x.title = NULL,
  y.title = "Variant Allele Frequency (\%)", out.file = NULL, width = 7,
  height = 5)
}
\description{
Plot the mean/median of the clusters of variants across samples
}

