% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.package.r
\name{plot.plant}
\alias{plot.plant}
\title{Plot method for plant objects.}
\usage{
\method{plot}{plant}(x, trunk.width = 20, add = FALSE,
  tree.rotation = 180, text.angle = NULL, branch.width = 1,
  branch.text.size = 0.5, node.size = 2, node.label.size = 0.75,
  node.text.size = 0.5, tree.label = NULL, ...)
}
\arguments{
\item{x}{The \code{plant} object to be plotted.}

\item{trunk.width}{The plotting line width for the trunk. Successive child
branches are plotted with increasingly finer \code{lwd}, to a minimum of
\code{1}.}

\item{add}{If \code{TRUE}, the plant will be added to the current plot.}

\item{...}{Additional arguments passed to \code{plot} and \code{segments}.}

\item{tree.rotation:}{tree rotation
values = c(0, 90, 180) ~ (bottom-up, left-right, top-down)}

\item{text.angle:}{text angle, if NULL then auto determine}
}
\value{
\code{NULL}
}
\description{
Plot method for plant objects.
}
\seealso{
\code{\link{germinate}}
}
