% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trees.r
\name{plot.tree.clone.as.branch}
\alias{plot.tree.clone.as.branch}
\title{Plot tree}
\usage{
\method{plot}{tree.clone.as.branch}(mt, angle = 15, branch.width = 1,
  branch.text.size = 0.3, node.size = 3, node.label.size = 0.75,
  node.text.size = 0.5, event.sep.char = ",", show.event = TRUE,
  tree.rotation = 0, text.angle = NULL, tree.label = NULL,
  branch.border.width = NULL, ...)
}
\description{
Plot tree
}

