% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{plot.tree}
\alias{plot.tree}
\title{Plot clonal evolution tree}
\usage{
\method{plot}{tree}(v, node.shape = "square", display = "tree",
  node.size = 50, cell.frac.ci = T, node.prefix.to.add = NULL,
  out.prefix = NULL, out.format = "graphml")
}
\arguments{
\item{v:}{clonal structure data frame as the output of enumerate.clones}
}
\description{
Plot a tree representing the clonal structure of a sample
Return the graph object (with node name prefixed with node.prefix.to.add)
}

