% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{rescale.vaf}
\alias{rescale.vaf}
\title{Rescale VAF of subclones s.t. total VAF must not exceed parent clone VAF}
\usage{
rescale.vaf(v, down.scale = 0.99)
}
\arguments{
\item{v:}{clonal structure data frame as output of enumerate.clones}
}
\description{
Rescale VAF of subclones s.t. total VAF must not exceed parent
clone VAF. When infered using statistical test, sometime the estimated
total mean/median VAFs of subclones > VAF of parent clones which makes
drawing difficult. This function rescale
}

