% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{save.clonevol.results}
\alias{save.clonevol.results}
\title{Save clonevol output to file}
\usage{
save.clonevol.results(x, out.prefix)
}
\arguments{
\item{x:}{Output of infer.clonal.models}

\item{out.prefix:}{Output prefix}
}
\description{
Save clonevol output to file
}
