% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{scale.sample.position}
\alias{scale.sample.position}
\title{Prepare x positions to layout samples in bell plots}
\usage{
\method{scale}{sample.position}(xstarts, xstops, plot.total.length = 7,
  evenly.distribute = T)
}
\arguments{
\item{xstarts:}{sample-named vector of x start positions}

\item{xstops:}{sample-named vector of x stop positions}

\item{total.length:}{total length of all sample in drawing}
}
\description{
Prepare x axis positions to layout samples. This enable
incorporation of clinical timeline (eg. diagnostic time, surgery time,
therapies, etc.) to be included in bell plots. This function will
scale and create a list of xstart position and length of samples such
that it maintain the positions provided in xstarts and xstops.
Return a list of xstarts and lengths to be used in draw.sample.clones
}

