% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{set.position}
\alias{set.position}
\title{Set vertical position of clones within a sample clonal polygon visualization}
\usage{
set.position(v)
}
\arguments{
\item{v:}{clonal structure data frame as output of enumerate.clones}
}
\description{
All subclones of a clone will be positioned next to each
other from the bottom of the polygon of the parent clone.
this function set the y.shift position depending on VAF
}

