% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stat.r
\name{subclonal.test}
\alias{subclonal.test}
\title{Subclonal test H0: mean_X1 + mean_X2 <= mean_X}
\usage{
subclonal.test(vaf.col.name, parent.cluster, sub.clusters = NULL,
  boot = NULL, cdf = NULL, min.cluster.vaf = 0, alpha = 0.05)
}
\arguments{
\item{boot:}{output of generate.boot(), if NULL, test using VAFs}

\item{vaf.col.name:}{name of the VAF column}

\item{parent.cluster:}{the cluster of mutations representing the parent
clone}

\item{sub.clusters:}{vector of sub clusters to test if they can be
all directly arise from the parent clone; if NULL, the parent.cluster
will be tested against 0 (this will help determine if this (parent)clone
represents with high enough cell frac to report)}

\item{cdf:}{Clonal data frame with VAF}
}
\description{
Return a list of p-value and confidence intervals of X-(X1+X2)
if p-value is small, reject H0, otherwise, not enough evidence to reject H0
}

