% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{trim.clone.trees}
\alias{trim.clone.trees}
\title{Reduce clone trees to core trees after excluding clones
that are present in only a single sample}
\usage{
trim.clone.trees(merged.trees, remove.sample.specific.clones = TRUE,
  samples = NULL, seeding.aware.tree.pruning = TRUE)
}
\arguments{
\item{merged.trees:}{output of infer.clonal.models()$matched$merged.trees
or any list of trees produced by infer.clonal.models()$models}

\item{remove.sample.specific.clones:}{remove clones that are found in
only one sample before comparing, default=T (used when reducing
merged trees. Set this to F to compare and reduce multiple samples
with the same tree to one sample}
}
\description{
Reduce clone trees produced by infer.clonal.models
to the core models, ie. ones that do not affect how clonal evolve
and branch across samples. This function will strip off the clones
that involve only one sample, any excluded nodes, and compare the trees
and keep only ones that are different. Return a list of merged.trees
}
