% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{write.tree}
\alias{write.tree}
\title{Write clonal evolution tree to file
TODO: do!}
\usage{
write.tree(v, out.file, out.format = "tabular")
}
\description{
Write clonal evolution tree to file
TODO: do!
}

