% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{draw.clone}
\alias{draw.clone}
\title{Draw a polygon representing a clone evolution, annotated with cluster label
and cellular fraction}
\usage{
draw.clone(x, y, wid=1, len=1, col='gray', label=NA, cell.frac=NA)
}
\arguments{
\item{x:}{x coordinate}

\item{y:}{y coordinate}

\item{wid:}{width of the polygon (representing cellular fraction)}

\item{len:}{length of the polygon}

\item{col:}{fill color of the polygon}

\item{label:}{name of the clone}

\item{cell.frac:}{cellular fraction of the clone}

\item{cell.frac.position:}{position for cell.frac =
c('top.left', 'top.right', 'top.mid',
'right.mid', 'right.top', 'right.bottom',
'side', 'top.out')}
}
\description{
Draw a polygon representing a clone, annotated with cluster
label and cellular fraction
}

