% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{estimate.clone.vaf}
\alias{estimate.clone.vaf}
\title{Estimate VAFs of clones/clusters from clonality analysis result}
\usage{
estimate.clone.vaf(v, cluster.col.name = "cluster", vaf.col.names = NULL,
  vaf.in.percent = TRUE, method = "median", ref.count.col.names = NULL,
  var.count.col.names = NULL, depth.col.names = NULL)
}
\arguments{
\item{v:}{data frame containing variants' VAF and clustering}

\item{cluster.col.name:}{name of the column that hold the cluster ID}

\item{vaf.col.names:}{names of the columns that hold VAFs}

\item{method:}{median or mean}
}
\description{
Estimate VAFs of clones/clusters by calculating the median of
the VAFs of all variants provided.
}
