% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol2fishplot.r
\name{generateFishplotInputs}
\alias{generateFishplotInputs}
\title{Generate fishplot ready data from clonevol clonal evolution models}
\usage{
generateFishplotInputs(results, rescale = T, samples = NULL)
}
\arguments{
\item{results:}{output from infer.clonal.models}

\item{rescale:}{rescale VAF such that no sum rule is violated (eg. parent
should have VAF >= sum of its chilren's VAF}

\item{samples:}{the names of the samples to be used in fishplot (this should
be a subset of vaf.col.names parameter provided to infer.clonal.models); The
order of samples provided will be prerseved in the fishplot}
}
\description{
Generate fishplot ready data from clonevol clonal evolution models
}
\examples{
\dontrun{
  samples = c('P', 'R')
  v = data.frame(cluster=c(1,1,2,2),P=c(50,40,0,0),R=c(50,60,20,20))
  x = infer.clonal.models(variants=v, vaf.col.names=samples)
  f = generateFishplotInputs(results=x)
  fishes = createFishPlotObjects(f)
  pdf('fish.pdf', width=8, height=5)
  for (i in 1:length(fishes)){
    fish = layoutClones(fishes[[i]])
    fish = setCol(fish,f$clonevol.clone.colors)
    fishPlot(fish,shape="spline", title.btm="Patient", cex.title=0.5,
           vlines=seq(1, length(samples)), vlab=samples, pad.left=0.5)
  }
  dev.off()
}
}

