% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{get.clonevol.colors}
\alias{get.clonevol.colors}
\title{Get the hex string of the preset colors optimized for plotting both
polygon plots and mutation scatter plots, etc.}
\usage{
get.clonevol.colors(num.colors, strong.color = F)
}
\description{
Get the hex string of the preset colors optimized for plotting both
polygon plots and mutation scatter plots, etc.
}

