% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonevol.r
\name{make.graph}
\alias{make.graph}
\title{Construct igraph object from clonal structures of a sample}
\usage{
make.graph(v, cell.frac.ci = TRUE, node.annotation = "clone",
  node.colors = NULL)
}
\description{
Construct igraph object from clonal structures of a sample
}
