% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{merge.samples}
\alias{merge.samples}
\title{Merge multi region samples into a meta sample}
\usage{
\method{merge}{samples}(x, samples, new.sample, new.sample.group,
  ref.cols = NULL, var.cols = NULL)
}
\description{
Merge multi region samples into a meta sample
}

