% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{plot.clonal.models}
\alias{plot.clonal.models}
\title{Plot evolution models (polygon plots and trees) for multi samples}
\usage{
\method{plot}{clonal.models}(models, out.dir, matched = NULL,
  variants = NULL, box.plot = FALSE, cluster.col.name = "cluster",
  scale.monoclonal.cell.frac = TRUE, width = NULL, height = NULL,
  text.size = 1, tree.node.shape = "circle", tree.node.size = 50,
  tree.node.text.size = 1, out.format = "png", resolution = 300,
  overwrite.output = FALSE, max.num.models.to.plot = 10,
  cell.frac.ci = FALSE, show.score = T, out.prefix = "model")
}
\arguments{
\item{models:}{list of model output from infer.clonal.models function}

\item{out.dir:}{output directory for the plots}

\item{matched:}{data frame of compatible models for multiple samples}

\item{scale.monoclonal.cell.frac:}{c(TRUE, FALSE); if TRUE, scale cellular
fraction in the plots (ie. cell fraction will be scaled by 1/purity =
1/max(VAF*2))}

\item{width:}{width of the plots (in), if NULL, automatically choose width}

\item{height:}{height of the plots (in), if NULL, automatically choose height}

\item{out.format:}{format of the plot files ('png', 'pdf', 'pdf.multi.files')}

\item{resolution:}{resolution of the PNG plot file}

\item{overwrite.output:}{if TRUE, overwrite output directory, default=FALSE}

\item{max.num.models.to.plot:}{max number of models to plot; default = 100
if NULL, unlimited}
}
\description{
Plot evolution models inferred by infer.clonal.models function
Two types of plots are supported: polygon plot and tree plot
}

