% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boxplot.r
\name{plot.pairwise}
\alias{plot.pairwise}
\title{Plot values of columns pairwise}
\usage{
\method{plot}{pairwise}(data, col.names = c(), suffix = "",
  group.col.name = "cluster", group.col.is.integer = TRUE, colors = NULL,
  shapes = NULL, show.legend.title = F, sharedCategoryColName = "",
  onePage = T, multiPages = F, xMin = 0, xMax = 100, yMin = 0,
  yMax = 100, xMinSmall = 0, xMaxSmall = 70, yMinSmall = 0,
  yMaxSmall = 70, show.none.zero.count = F, out.prefix = "")
}
\arguments{
\item{data:}{a data frame}

\item{col.names:}{the columns to plot all pairwise, eg. VAFs of the samples}

\item{group.col.name:}{column used to determine the category
when plotting to give different shapes,colors. eg. the cluster identity
of the variants}
}
\description{
Plot values of columns pairwise
}

