% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{scale.cell.frac}
\alias{scale.cell.frac}
\title{Scale cellular fraction in a clonal architecture model}
\usage{
\method{scale}{cell.frac}(m, ignore.clusters = NULL)
}
\description{
Root clone VAF will be determined, and all vaf will be scaled such that
roo clone VAF will be 0.5
}

