% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clonevol.r
\name{transfer.events.to.consensus.trees}
\alias{transfer.events.to.consensus.trees}
\title{Transfer driver variants/events from the cluster onto the clonal
evolution tree, so the tree can be plot with events on branch}
\usage{
transfer.events.to.consensus.trees(x, events, cluster.col.name = "cluster",
  event.col.name)
}
\description{
Transfer driver variants/events from the cluster onto the clonal
evolution tree, so the tree can be plot with events on branch
}

